# server/docs/swagger.yaml
basePath: /v1
definitions:
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
...
  queries.Card:
    properties:
      art:
        type: string
      id:
        type: integer
      suit:
        $ref: '#/definitions/queries.Cardsuit'
      value:
        $ref: '#/definitions/queries.Cardvalue'
    type: object
  queries.Cardstate:
    enum:
    - Deck
    - Hand
    - Play
    - Kitty
    type: string
    x-enum-varnames:
    - CardstateDeck
    - CardstateHand
    - CardstatePlay
    - CardstateKitty
  queries.Cardsuit:
    enum:
    - Spades
    - Clubs
    - Hearts
    - Diamonds
    type: string
    x-enum-varnames:
    - CardsuitSpades
    - CardsuitClubs
    - CardsuitHearts
    - CardsuitDiamonds
  queries.Cardvalue:
    enum:
    - Ace
    - Two
    - Three
    - Four
    - Five
    - Six
    - Seven
    - Eight
    - Nine
    - Ten
    - Jack
    - Queen
    - King
    - Joker
    type: string
    x-enum-varnames:
    - CardvalueAce
    - CardvalueTwo
    - CardvalueThree
    - CardvalueFour
    - CardvalueFive
    - CardvalueSix
    - CardvalueSeven
    - CardvalueEight
    - CardvalueNine
    - CardvalueTen
    - CardvalueJack
    - CardvalueQueen
    - CardvalueKing
    - CardvalueJoker
  queries.Deck:
    properties:
      cutmatchcardid:
        $ref: '#/definitions/pgtype.Int4'
      id:
        type: integer
    type: object
  queries.DeckMatchcard:
    properties:
      deckid:
        type: integer
      matchcardid:
        type: integer
    type: object
  queries.Gamestate:
    enum:
    - NewGameState
    - JoinGameState
    - WaitingState
    - MatchReady
    - DealState
    - DiscardState
    - CutState
    - PlayState
    - OpponentState
    - KittyState
    - GameWonState
    - GameLostState
    - MaxGameState
    type: string
    x-enum-varnames:
    - GamestateNewGameState
    - GamestateJoinGameState
    - GamestateWaitingState
    - GamestateMatchReady
    - GamestateDealState
    - GamestateDiscardState
    - GamestateCutState
    - GamestatePlayState
    - GamestateOpponentState
    - GamestateKittyState
    - GamestateGameWonState
    - GamestateGameLostState
    - GamestateMaxGameState
  queries.GetMatchCardsRow:
    properties:
      card:
        $ref: '#/definitions/queries.Card'
      deck:
        $ref: '#/definitions/queries.Deck'
      deckMatchcard:
        $ref: '#/definitions/queries.DeckMatchcard'
      matchcard:
        $ref: '#/definitions/queries.Matchcard'
    type: object
  queries.Match:
    properties:
      art:
        type: string
      creationdate:
        $ref: '#/definitions/pgtype.Timestamptz'
      currentplayerturn:
        type: integer
      cutgamecardid:
        type: integer
      deckid:
        type: integer
      elorangemax:
        type: integer
      elorangemin:
        type: integer
      gamestate:
        $ref: '#/definitions/queries.Gamestate'
      id:
        type: integer
      privatematch:
        type: boolean
      turnpasstimestamps:
        items:
          $ref: '#/definitions/pgtype.Timestamptz'
        type: array
    type: object
  queries.Matchcard:
    properties:
      cardid:
        type: integer
      currowner:
        $ref: '#/definitions/pgtype.Int4'
      id:
        type: integer
      origowner:
        $ref: '#/definitions/pgtype.Int4'
      state:
        $ref: '#/definitions/queries.Cardstate'
    type: object
...
  vo.JoinMatchReq:
    properties:
      matchId:
        type: integer
      playerId:
        type: integer
    type: object
  vo.MatchDetailsResponse:
    properties:
      gameState:
        $ref: '#/definitions/queries.Gamestate'
      matchId:
        type: integer
      playerId:
        type: integer
    type: object
  vo.MatchRequirements:
    properties:
      accountId:
        type: integer
      eloRangeMax:
        type: integer
      eloRangeMin:
        type: integer
      isPrivate:
        type: boolean
      playerId:
        type: integer
    type: object
host: localhost:1323
info:
  contact: {}
  description: cribbage rest server
  title: cribbage server
  version: 0.0.2
paths:
...
  /player/match/:
    get:
      consumes:
      - application/json
      parameters:
      - description: search for match by id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.MatchDetailsResponse'
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
      summary: Get match by id
      tags:
      - match
    post:
      consumes:
      - application/json
      parameters:
      - description: MatchRequirements
        in: body
        name: details
        required: true
        schema:
          $ref: '#/definitions/vo.MatchRequirements'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create new match
      tags:
      - match
    put:
      consumes:
      - application/json
      parameters:
      - description: match Object to update
        in: body
        name: details
        required: true
        schema:
          $ref: '#/definitions/queries.Match'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queries.Match'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update match by id
      tags:
      - match
  /player/match/cards/:
    get:
      consumes:
      - application/json
      parameters:
      - description: search for match by id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/queries.GetMatchCardsRow'
            type: array
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
      summary: Get match cards by id
      tags:
      - match
...
  /player/match/join:
    put:
      consumes:
      - application/json
      parameters:
      - description: match Object to update
        in: body
        name: details
        required: true
        schema:
          $ref: '#/definitions/vo.JoinMatchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.MatchDetailsResponse'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Join match by id
      tags:
      - match
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"